pyproject.toml
[project]
name = "cuttle-simulation"
version = "0.1.0"
description = "Cuttle card game simulator for strategy research"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "hypothesis>=6.0",
    "ruff>=0.1",
    "mypy>=1.0",
]
llm = [
    "anthropic>=0.30",
    "openai>=1.0",
    "httpx>=0.27",
]
api = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "python-dotenv>=1.0",
]
cloud = [
    "modal>=0.60",
    "redis>=5.0",
    "asyncpg>=0.29",
    "clickhouse-connect>=0.7",
]

[project.scripts]
cuttle = "cuttle_engine.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cuttle_engine", "strategies", "simulation", "analytics", "api", "web", "db", "core"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
